@@ -9,7 +9,6 @@
   "scripts": {
     "start": "node dist/src/server.js",
     "dev": "tsnd --transpile-only --ignore-watch node_modules --respawn src/server.ts",
-    "test": "NODE_ENV=TEST jest",
     "postinstall": "tsc",
     "watch-node": "tsnd --transpile-only --ignore-watch node_modules --respawn src/server.ts",
     "watch-ts": "tsc -w",
@@ -40,9 +39,11 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.13"
+    "mongoose": "~6.6.0"
   },
   "devDependencies": {
+    "@typegoose/typegoose": "^9.12.1",
+    "@types/bcrypt": "^5.0.0",
     "@types/cors": "^2.8.8",
     "@types/express": "^4.17.9",
     "@types/jest": "^26.0.15",
@@ -56,8 +57,8 @@
     "jest": "^26.6.3",
     "supertest": "^6.0.1",
     "ts-jest": "^26.4.4",
-    "ts-node-dev": "^1.0.0",
-    "typescript": "^4.0.5"
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^4.4"
   },
   "optionalDependencies": {
     "win-node-env": "^0.4.0"
