@@ -1,16 +1,10 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
+
 const connectString =
   process.env.NODE_ENV === "TEST"
     ? `mongodb://localhost/${Date.now()}`
     : process.env.MONGODB_URI;
 
-mongoose.connect(connectString, {
-  useNewUrlParser: true,
-  useCreateIndex: true,
-  useUnifiedTopology: true,
-  useFindAndModify: false,
-});
+const conn = mongoose.createConnection(connectString);
 
-module.exports = mongoose;
-
-export {};
+export default conn;
