@@ -35,6 +35,8 @@ module.exports = {
       "always"
     ],
     "@typescript-eslint/no-var-requires": 0,
-    "@typescript-eslint/no-explicit-any": 0
+    "@typescript-eslint/no-explicit-any": 0,
+    "@typescript-eslint/ban-types": 0,
+    "@typescript-eslint/no-non-null-assetion": 0,
   }
 };
