@@ -6,8 +6,8 @@ require("dotenv").config({
 });
 const cors = require("cors");
 
-const rotasFaturas = require("./app/routes/faturas");
-const rotasUsuarios = require("./app/routes/users");
+
+const usersRoutes = require("./app/routes/users");
 
 const corsConfig = require("./config/cors");
 
@@ -20,8 +20,8 @@ app.use(corsConfig);
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-app.use("/users", rotasUsuarios);
-app.use("/faturas", rotasFaturas);
+app.use("/users", usersRoutes);
+
 app.use((req: Request, res: Response) => {
   res.status(404).json({ 
     error: "Invalid endpoint.",
